scripts/zim/requirements.txt
dateutils
feedparser
